backend/requirements.txt
django==4.2.7

djangorestframework==3.14.0

django-cors-headers==4.3.1


flake8==6.1.0
isort==5.12.0

psycopg2==2.9.9

django-bootstrap-datepicker-plus==5.0.4

Pillow==10.1.0

stripe==7.13.0
